pyproject.toml
[project]
name = "ppc-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "google-ads==27.0.0",
    "google-generativeai==0.8.3",
    "gradio<5.9.0",
    "pandas==2.2.3",
    "pydantic==2.9.0",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "uvicorn[standard]==0.30.6",
]
